[DEV-6440] move validate_fiscal_period
