+++ usaspending_api/reporting/v2/views/submission_history.py
@@ -2,7 +2,6 @@ from rest_framework.response import Response
 
 from usaspending_api.agency.v2.views.agency_base import AgencyBase, PaginationMixin
 
-from usaspending_api.common.exceptions import UnprocessableEntityException
 from usaspending_api.common.helpers.generic_helper import get_pagination_metadata
 from usaspending_api.submissions.models import SubmissionAttributes
 
@@ -12,12 +11,6 @@ class SubmissionHistory(PaginationMixin, AgencyBase):
 
     endpoint_doc = "usaspending_api/api_contracts/contracts/v2/reporting/agencies/agency_code/fiscal_year/fiscal_period/submission_history.md"
 
-    @staticmethod
-    def validate_fiscal_period(fiscal_period):
-        fiscal_period = int(fiscal_period)
-        if fiscal_period < 2 or fiscal_period > 12:
-            raise UnprocessableEntityException(f"fiscal_period must be in the range 2-12")
-
     def get(self, request, toptier_code, fiscal_year, fiscal_period):
         self.fiscal_year = int(fiscal_year)
         self.sortable_columns = [
@@ -25,7 +18,7 @@ class SubmissionHistory(PaginationMixin, AgencyBase):
             "certified_date",
         ]
         self.default_sort_column = "publication_date"
-        self.validate_fiscal_period(fiscal_period)
+        self.validate_fiscal_period({"fiscal_period": fiscal_period})
         results = (
             SubmissionAttributes.objects.filter(
                 toptier_code=toptier_code, reporting_fiscal_year=fiscal_year, reporting_fiscal_period=fiscal_period,
